Upgraded to the latest Kotlin 1.7.x
Applied a workaround for the compilation issue by adding an explicit
cast.

+++ app/src/androidTest/java/org/northwinds/app/sotachaser/ui/summits/SummitFragmentTest.kt
@@ -264,7 +264,7 @@ class SummitFragmentTest {
             if (view == null)
                 throw noMatchingViewException
             val recyclerView = view as RecyclerView
-            assertThat(recyclerView.adapter?.itemCount,
+            assertThat(recyclerView.adapter?.itemCount as Int,
                 Matchers.`is`(Matchers.greaterThanOrEqualTo(10))
             )
         }
